Manila: Put installation specific tasks in a stand-alone module
